Addressing comments from reviews
