+++ TwoPageEBookReader/README.md
@@ -1,6 +1,6 @@
 # eBook Reader sample using Window Manager
 
-This sample demonstrates the Two Page design pattern for foldable devices. In this pattern, apps can take advantage of a foldable device's natural boundary to show multiple items from a collection. EBooks are a perfect application of this pattern, as their physical predecessors could be called "dual-leaf hinged devices". See other design patterns in [Microsoft's UX Design documentation](https://docs.microsoft.com/en-us/dual-screen/introduction).
+This sample demonstrates the Two Page design pattern for foldable devices. In this pattern, apps can take advantage of a foldable device's natural boundary to show multiple items from a collection. EBooks are a perfect application of this pattern, as their physical predecessors could be called "dual-leaf hinged devices". See other design patterns in [Microsoft's UX Design documentation](https://docs.microsoft.com/dual-screen/introduction).
 
 ![A dual-leaf hinged device](screenshots/base.png)
 
@@ -20,3 +20,11 @@ Very little extra code in `BookActivity` was necessary to enable two-page layout
 In `StartActivity`, where users choose a book to read, Android Jetpack's [SlidingPaneLayout](https://developer.android.com/reference/androidx/slidingpanelayout/widget/SlidingPaneLayout) can keep the book chooser on one side of a fold. There are definitely better, more in-depth samples demonstrating `SlidingPaneLayout` in public Microsoft repositories. However, the solution for this application was simple: include a [SlidingPaneLayout](app/src/main/res/layout/activity_start.xml) with functional elements on the left and an empty, 0-width `View` on the right. *This trick required only a few lines of XML, and zero code.*
 
 ![StartActivity](screenshots/start.png)
+
+## Project Gutenberg Attributions
+
+The following license can be found for every Project Gutenberg ebook in this app, in the Attributions section.
+
+> This eBook is for the use of anyone anywhere in the United States and most other parts of the world at no cost and with almost no restrictions whatsoever. You may copy it, give it away or re-use it under the terms of the Project Gutenberg License included with this eBook or online at www.gutenberg.org. If you are not located in the United States, you will have to check the laws of the country where you are located before using this eBook.
+
+> Find the full Project Gutenberg license in the last chapter. Original formats of this text by Project Gutenberg can be found in the app/src/main/assets/ folder, as well as [www.gutenberg.org](https://www.gutenberg.org/)
