+++ TwoPageEBookReader/app/src/main/java/com/microsoft/device/wm_samples/ebook_reader_sample/BookPager.kt
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License.
  */
-package com.example.ebook_reader_sample
+package com.microsoft.device.wm_samples.ebook_reader_sample
 
 import android.text.Html
 import android.view.LayoutInflater
